🐛 fix : [관심종목] 종목검색 시 최근검색종목 추가 랜더링되도록 변경
