+++ src/app/(afterLogin)/favorite/_components/FavoriteTitleSection.tsx
@@ -4,7 +4,6 @@ import TextButton from "@/common/TextButton";
 import React, { useEffect, useState } from "react";
 import AddFavoriteModal from "./AddFavoriteModal";
 import { TstockInfoList } from "./FavoriteStockLists";
-import { stockListByStockName } from "@/hooks/discovery/useSearchAction";
 import { useRecentKeywordStore, TrecentData } from "@/store/useRecentKeywordStore";
 
 export type TFavoriteTitleSection = {
@@ -18,33 +17,6 @@ export type TFavoriteTitleSection = {
 export default function FavoriteTitleSection({ popularSearchData }: TFavoriteTitleSection) {
   const [isAddFavoriteModalOpened, setAddFavoriteModalOpened] = useState(false); // 모달 열고닫기
   const [stockInfoList, setStockInfoList] = useState<TstockInfoList>([]);
-  const { recentKeywordList, setRecentKeywordList } = useRecentKeywordStore(); //
-
-  // 최근 검색종목 데이터가져오기
-  useEffect(() => {
-    (async () => {
-      const savedRecentData = localStorage.getItem("recentData");
-      if (savedRecentData) {
-        const parsedRecentData: TrecentData = JSON.parse(savedRecentData);
-        setRecentKeywordList(parsedRecentData);
-        console.log("최근검색어", parsedRecentData);
-
-        const stockNameList = parsedRecentData.map((item) => item.keyword);
-        if (stockNameList.length > 0) {
-          try {
-            const recentStockDataList = await stockListByStockName(stockNameList);
-            console.log("최근검색데이터리스트", recentStockDataList);
-            if (recentStockDataList) {
-              setStockInfoList(recentStockDataList);
-              console.log("최근검색어의 주식데이터", recentStockDataList[0]);
-            }
-          } catch (e) {
-            console.error("Error:", e);
-          }
-        }
-      }
-    })();
-  }, [setRecentKeywordList]);
 
   return (
     <>
@@ -66,11 +38,7 @@ export default function FavoriteTitleSection({ popularSearchData }: TFavoriteTit
         </div>
       </div>
       {isAddFavoriteModalOpened && (
-        <AddFavoriteModal
-          onClose={() => setAddFavoriteModalOpened(false)}
-          popularSearchData={popularSearchData}
-          recentData={stockInfoList}
-        />
+        <AddFavoriteModal onClose={() => setAddFavoriteModalOpened(false)} popularSearchData={popularSearchData} />
       )}
     </>
   );
